JS-7715: logo, language, arrow back
